Input form Update React thinking
