--- pages/test.tsx
@@ -1,10 +1,0 @@
-import { NextPage } from "next";
-import SearchByName from "../components/dynamic/SearchByName.component";
-
-const TestPage: NextPage = () => {
-    return <div>
-        <SearchByName />
-    </div>
-}
-
-export default TestPage;
